requirements.txt
# Third-party packages
requests>=2.31.0
pandas>=2.0.0
openpyxl

# Google Cloud Platform
google-auth>=2.20.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.10.0
google-api-python-client>=2.80.0